Added noise to InfluxDB timestamps.
Previous version indexed the tweet ID as a tag to prevent overwrites, as the Twitter API produces second-level timestamps. We don't really use the ID as an index so I decided to jitter the timestamp a little bit by converting it to microseconds and adding random noise. This prevents duplicates and avoids indexing the Tweet ID.